backtracking warning implemented
